requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
boto3>=1.35.0
chromadb>=0.4.0
python-dotenv>=1.0.0
